+++ fixtures/ext-types/guid/tests/bindings/test_guid.py
@@ -16,6 +16,30 @@ class TestGuid(unittest.TestCase):
         self.assertEqual(helper.guids, ["second-guid", "third-guid"])
         self.assertEqual(helper.maybe_guid, None)
 
+    def test_get_guid_errors(self):
+        # This is testing `get_guid` which never returns a result, so everything
+        # is InternalError representing a panic.
+        # The fixture hard-codes some Guid strings to return specific errors.
+        with self.assertRaisesRegex(InternalError, "Failed to convert arg 'val': The Guid is too short"):
+            get_guid("")
+
+        with self.assertRaisesRegex(InternalError, "Failed to convert arg 'val': Something unexpected went wrong"):
+            get_guid("unexpected")
+
+        with self.assertRaisesRegex(InternalError, "guid value caused a panic!"):
+            get_guid("panic")
+
+    def test_try_get_guid_errors(self):
+        # This is testing `try_get_guid()` which says it returns a result, so we
+        # will get a mix of "expected" errors and panics.
+        with self.assertRaises(GuidError.TooShort):
+            try_get_guid("")
+
+        with self.assertRaisesRegex(InternalError, "Failed to convert arg 'val': Something unexpected went wrong"):
+            try_get_guid("unexpected")
+
+        with self.assertRaisesRegex(InternalError, "guid value caused a panic!"):
+            try_get_guid("panic")
 
     # def test_round_trip(self):
     #     ct = get_combined_type(None)